# Optimize zed.Value.Under