Move the loop in Value.Under into a separate method, Value.under, so the compiler will inline Under.

goos: darwin
goarch: arm64
pkg: github.com/brimdata/zed
                        |  before.txt  |              after.txt               |
                        |    sec/op    |    sec/op     vs base                |
ValueUnder/primitive-10   2.0560n ± 2%   0.9328n ± 0%  -54.63% (p=0.000 n=10)
ValueUnder/named-10        27.77n ± 0%    27.96n ± 1%   +0.72% (p=0.000 n=10)
geomean                    7.555n         5.107n       -32.40%